+++ src/main/java/org/example/broong/config/JwtUtil.java
@@ -12,7 +12,7 @@ import java.util.Base64;
 import java.util.Date;
 import lombok.extern.slf4j.Slf4j;
 import org.example.broong.global.exception.ApiException;
-import org.example.broong.user.enums.UserType;
+import org.example.broong.domain.user.enums.UserType;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
