+++ src/main/java/org/example/broong/domain/auto/controller/AutoController.java
@@ -1,10 +1,10 @@
-package org.example.broong.auto.controller;
+package org.example.broong.domain.auto.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.example.broong.auto.dto.request.SignupRequestDto;
-import org.example.broong.auto.dto.response.SignupResponseDto;
-import org.example.broong.auto.service.AutoService;
+import org.example.broong.domain.auto.dto.request.SignupRequestDto;
+import org.example.broong.domain.auto.dto.response.SignupResponseDto;
+import org.example.broong.domain.auto.service.AutoService;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -18,10 +18,9 @@ public class AutoController {
     private final AutoService autoService;
 
     @PostMapping("/signup")
-    public SignupResponseDto signup(@Valid @RequestBody SignupRequestDto requestDto){
+    public SignupResponseDto signup(@Valid @RequestBody SignupRequestDto requestDto) {
         return autoService.signup(requestDto);
     }
-
 
 
 }
