+++ src/main/java/org/example/broong/domain/auto/dto/request/SignupRequestDto.java
@@ -1,4 +1,4 @@
-package org.example.broong.auto.dto.request;
+package org.example.broong.domain.auto.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.example.broong.user.enums.UserType;
+import org.example.broong.domain.user.enums.UserType;
 
 @Getter
 @NoArgsConstructor
@@ -21,10 +21,10 @@ public class SignupRequestDto {
     private String email;
 
     @NotBlank
-    @Pattern(regexp = "^(?=.*[A-Z])(?=.*\\d)[a-zA-Z0-9]{8,}$" , message = "새 비밀번호는 8자 이상이어야 하고, 숫자와 대문자를 포함해야 합니다.")
+    @Pattern(regexp = "^(?=.*[A-Z])(?=.*\\d)[a-zA-Z0-9]{8,50}$" , message = "새 비밀번호는 8자 이상이어야 하고, 숫자와 대문자를 포함해야 합니다.")
     private String password;
 
     @NotBlank
-    private String userType;
+    private UserType userType;
 
 }
