+++ src/main/java/org/example/broong/domain/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package org.example.broong.user.repository;
+package org.example.broong.domain.user.repository;
 
-import org.example.broong.user.entity.User;
+import org.example.broong.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long>{
